task1-Knowledge-Retrieval/requirements.txt
# Core Framework
streamlit>=1.28.0

# Azure Services
azure-search-documents>=11.4.0
azure-identity>=1.15.0
openai>=1.12.0

# PDF Processing
pypdf>=3.17.0
pdfplumber>=0.10.3

# Text Processing
tiktoken>=0.5.2
langchain-text-splitters>=0.0.1

# Configuration
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.3
